requirements.txt
langchain==0.1.0
langchain-groq==0.1.0
langchain-community==0.0.12
langsmith==0.0.87
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
chromadb==0.4.22
groq==0.4.1